requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
gunicorn==21.2.0
twilio==9.8.4
python-dotenv==1.1.1
sqlalchemy==2.0.44
google-generativeai==0.8.5
jinja2==3.1.6
aiofiles==24.1.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
python-jose[cryptography]==3.3.0
chromadb==0.4.22
langchain==0.1.0
langchain-google-genai==0.0.6